requirements.txt
# Core dependencies for CPU-only development
streamlit
sentence-transformers
langchain
pyvis
faiss-cpu
PyPDF2
pdfplumber
PyMuPDF
python-docx
numpy
langchain-text-splitters
python-dotenv
google-generativeai
Pillow